requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
pillow
kaggle
torch
torchvision
torchaudio
pyinstaller
twine
build
google-genai
python-dotenv
pypdf
python-docx
playwright
pydantic
mcp